+++ SLUGS/urls.py
@@ -25,9 +25,7 @@ urlpatterns = [
     path("unicorn/", include("django_unicorn.urls")),
     path("tinymce/", include("tinymce.urls")),
     path("__debug__/", include(debug_toolbar.urls)),
-    path('hijack/', include('hijack.urls')),
-
-    
+    path("hijack/", include("hijack.urls")),
     path("admin/", admin.site.urls),
     path(
         "media/forms/<path:relative_path>", FormDownload.as_view(), name="download_form"