+++ employee/views.py
@@ -26,7 +26,7 @@ from dev_utils.views import MultipleFormView
 from SLUGS.views import SLUGSMixin, isAdminMixin
 from utils.models import onboardingStatus
 from finance.utils import getShiftsForEmployee
-from finance.forms import ShiftFormSet
+from finance.forms import OfficeHoursShiftFormSet
 from finance.models import Shift
 from employee.models import OfficeHours, PaperworkForm
 
@@ -83,17 +83,21 @@ class userOverview(SLUGSMixin, MultipleFormView):
         self.added_context["timeworked"] = self.added_context["shifts"].aggregate(
             Sum("total_time")
         )
-        self.added_context["processed_shifts"] = self.added_context["shifts"].filter(
-            processed=True
-        )[:100]
+        self.added_context["processed_shifts"] = (
+            self.added_context["shifts"]
+            .order_by("-time_out")
+            .filter(processed=True)[:500]
+        )
         self.added_context["amount_made"] = self.added_context[
             "processed_shifts"
         ].aggregate(Sum("cost"))
-        self.added_context["unprocessed_shifts"] = self.added_context["shifts"].filter(
-            processed=False, contested=False
+        self.added_context["unprocessed_shifts"] = (
+            self.added_context["shifts"]
+            .order_by("-time_out")
+            .filter(processed=False, contested=False)
         )
-        self.added_context["contested_shifts"] = self.added_context["shifts"].filter(
-            contested=True
+        self.added_context["contested_shifts"] = (
+            self.added_context["shifts"].order_by("-time_out").filter(contested=True)
         )
         return super().dispatch(request, *args, **kwargs)
 
@@ -111,25 +115,25 @@ class officeHours(SLUGSMixin, MultipleFormView):
         except Exception:
             pass
         for office_hour_obj in OfficeHours.objects.filter(employee=request.user):
-            self.added_context['form_meta'][f"office_hours_{office_hour_obj.pk}"] = {
+            self.added_context["form_meta"][f"office_hours_{office_hour_obj.pk}"] = {
                 "obj": office_hour_obj,
-                "rate": office_hour_obj.position
+                "rate": office_hour_obj.position,
             }
             self.form_classes[f"office_hours_{office_hour_obj.pk}"] = {
-                "form": ShiftFormSet,
+                "form": OfficeHoursShiftFormSet,
                 "kwargs": {
                     "queryset": Shift.objects.filter(
                         object_id=office_hour_obj.id,
                         content_type_id=ContentType.objects.get(model="officehours").id,
                         processed=False,
                     ),
-                }
+                },
             }
         return super().dispatch(request, *args, **kwargs)
 
     def process_forms(self, form_instances):
         for form in form_instances["forms"]:
-            office_hour_obj = self.added_context['form_meta'][form]['obj']
+            office_hour_obj = self.added_context["form_meta"][form]["obj"]
             for shift in form_instances["forms"][form].forms:
                 if (
                     shift.instance.time_in is not None
@@ -228,6 +232,7 @@ class massAssignPaperwork(SLUGSMixin, FormView):
         )
         return super().form_valid(form)
 
+
 class addGroups(SLUGSMixin, isAdminMixin, FormView):
     form_class = addGroupsForm
     template_name = "employee/add_groups.html"
@@ -240,7 +245,5 @@ class addGroups(SLUGSMixin, isAdminMixin, FormView):
 
     def form_valid(self, form):
         form.add_groups(self.request)
-        messages.add_message(
-            self.request, messages.SUCCESS, f"Groups added"
-        )
+        messages.add_message(self.request, messages.SUCCESS, f"Groups added")
         return super().form_valid(form)