+++ finance/views.py
@@ -26,6 +26,7 @@ class viewEstimate(SLUGSMixin, TemplateView):
     added_context = {"systems": {}, "fees": {}}
 
     def dispatch(self, request, *args, **kwargs):
+        Estimate.objects.get(pk=kwargs["e_id"]).save()
         self.added_context = calculateGigCost(Estimate.objects.get(pk=kwargs["e_id"]))
         return super().dispatch(request, *args, **kwargs)
 
@@ -191,7 +192,11 @@ class exportSummaryCSV(SLUGSMixin, View):
             ["", ""]
             + ["" for rate in sumData["rates"]]
             + ["", ""]
-            + [sumData["pay_period"].start, sumData["pay_period"].end, sumData["pay_period"].payday]
+            + [
+                sumData["pay_period"].start,
+                sumData["pay_period"].end,
+                sumData["pay_period"].payday,
+            ]
         )
         for emp in sumData["employees"]:
             emp = sumData["employees"][emp]